[fit]: 리뷰 삭제 시 멱등성 유지 삭제를 위한 조회 시 데이터가 없어도 예외 발생하지 않게 수정 및 사용자 검증 추가

+++ BE/backend/src/main/java/project/main/webstore/domain/review/service/ReviewService.java
@@ -92,14 +92,13 @@ public class ReviewService {
         Optional.ofNullable(review.getComment()).ifPresent((findReview::setComment));
 
         //이미지 수정하는 로직
-        if (imageInfo != null) {
-            Image image = imageUtils.patchImageWithDelete(imageInfo, findReview.getReviewImage());
-            ReviewImage reviewImage = createReviewImage(image, findReview);
-            findReview.addOrDeleteReviewImage(reviewImage);
-        }
+        Image image = imageUtils.patchImageWithDelete(imageInfo, findReview.getReviewImage());
+        ReviewImage reviewImage = createReviewImage(image, findReview);
+        findReview.addOrDeleteReviewImage(reviewImage);
 
         return findReview;
     }
+
     public Review patchReview(Review review) {
         Review findReview = reviewValidService.validReview(review.getId());
 
@@ -119,11 +118,16 @@ public class ReviewService {
         return findReview;
     }
 
-    public void deleteReview(Long reviewId) {
-        Review findReview = reviewValidService.validReview(reviewId);
-        ReviewImage reviewImage = findReview.getReviewImage();
-        imageUtils.deleteImage(reviewImage);
-        reviewRepository.deleteById(reviewId);
+    public void deleteReview(Long reviewId,Long userId) {
+        Optional<Review> findReview = reviewRepository.findById(reviewId);
+        if(findReview.isPresent()){
+            Review review = findReview.get();
+
+            userValidService.validUserIdSame(userId, review);
+
+            imageUtils.deleteImage(review.getReviewImage());
+            reviewRepository.deleteById(reviewId);
+        }
     }
 
     public List<Review> bestReviewByAdmin(List<Long> reviewIdList) {
@@ -139,7 +143,7 @@ public class ReviewService {
         return reviewList.stream().filter(review -> review.isBest()).collect(Collectors.toList());
     }
 
-    private ReviewImage createReviewImage(Image image, Review review){
+    private ReviewImage createReviewImage(Image image, Review review) {
         return image != null ? new ReviewImage(image, review) : null;
     }
 }
